# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Basic API",
        "description": "L5 Swagger API with Laravel basic",
        "contact": {
            "email": "nguyen.an.thoi.212@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/laravel_basic/an.thoi.laravel/public",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create a new user",
                "description": "",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Create a user",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "default": "admin_123@gmail.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "default": "84 123456789"
                                },
                                "password": {
                                    "type": "string",
                                    "default": "123456"
                                },
                                "name": {
                                    "type": "string",
                                    "default": "admin 123"
                                },
                                "gender": {
                                    "type": "boolean",
                                    "default": false
                                },
                                "dateOfBirth": {
                                    "type": "string",
                                    "default": "10/10/2010"
                                },
                                "about": {
                                    "type": "string",
                                    "default": "T la admin"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "user login",
                        "required": true,
                        "schema": {
                            "properties": {
                                "emailOrPhone": {
                                    "type": "string",
                                    "default": "admin_123@gmail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "default": "123456"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/request-reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request Reset Password",
                "description": "Email Request Reset Password",
                "operationId": "ResetPassword",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "default": "admin_123@gmail.com"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request Reset Password Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/accept-reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Accept Reset Password",
                "description": "Accept Token Reset Password",
                "operationId": "postTokenResetPassword",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "input password to reset",
                        "required": true,
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string",
                                    "default": "123456"
                                },
                                "password": {
                                    "type": "string",
                                    "default": "123456"
                                },
                                "confirmPassword": {
                                    "type": "string",
                                    "default": "123456"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset Password Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/change-password": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "description": "",
                "operationId": "ChangePassword",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "properties": {
                                "currentPassword": {
                                    "type": "string",
                                    "default": "123456789"
                                },
                                "newPassword": {
                                    "type": "string",
                                    "default": "123456"
                                },
                                "confirmNewPassword": {
                                    "type": "string",
                                    "default": "123456"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Change Password Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "A short description for security scheme",
            "name": "Authorization",
            "in": "header"
        }
    }
}